src/Module.Main/requirements.txt
coreschema==0.0.4
env-file==2020.12.3
python-decouple==3.4
absl-py==0.11.0
astunparse==1.6.3
cachetools==4.2.0
flatbuffers==1.12
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
Keras-Preprocessing==1.1.2
Markdown==3.3.3
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
python-dateutil==2.8.1
requests-oauthlib==1.3.0
rsa==4.7
seqlog==0.3.19
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
tensorflow-serving-api==2.4.0
termcolor==1.1.0
typing-extensions==3.7.4.3
Werkzeug==1.0.1
tabulate==0.8.7
protobuf==3.9.2